Livestream Dec 17 2021

+++ CwkSocial/CwkSocial.Application/UserProfiles/CommandHandlers/DeleteUserProfileHandler.cs
@@ -8,25 +8,39 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Cwk.Domain.Aggregates.UserProfileAggregate;
+using CwkSocial.Application.Enums;
+using CwkSocial.Application.Models;
 
 namespace CwkSocial.Application.UserProfiles.CommandHandlers
 {
-    internal class DeleteUserProfileHandler : IRequestHandler<DeleteUserProfile>
+    internal class DeleteUserProfileHandler : IRequestHandler<DeleteUserProfile, OperationResult<UserProfile>>
     {
         private readonly DataContext _ctx;
         public DeleteUserProfileHandler(DataContext ctx)
         {
             _ctx = ctx;
         }
-        public async Task<Unit> Handle(DeleteUserProfile request, CancellationToken cancellationToken)
+        public async Task<OperationResult<UserProfile>> Handle(DeleteUserProfile request, CancellationToken cancellationToken)
         {
+            var result = new OperationResult<UserProfile>();
             var userProfile = await _ctx.UserProfiles
                 .FirstOrDefaultAsync(up => up.UserProfileId == request.UserProfileId);
+            
+            if (userProfile is null)
+            {
+                result.IsError = true;
+                var error = new Error { Code = ErrorCode.NotFound, 
+                    Message = $"No UserProfile found with ID {request.UserProfileId}"};
+                result.Errors.Add(error);
+                return result;
+            }
 
             _ctx.UserProfiles.Remove(userProfile);
             await _ctx.SaveChangesAsync();
 
-            return new Unit();
+            result.Payload = userProfile;
+            return result;
         }
     }
 }
